Fix cname test case and add new test + Fix CNAME handling
Search for 'cname.'+DOMAIN since that's the CNAME record.
Add another cname test, test_find_domain_dual_cnames, which
tests that the first CNAME is skipped and the second CNAME
is found.

A CNAME is unlike a DNAME since a CNAME is an alias for one name
not for a subtree. Continue the search if the CNAME record isn't
the complete domain name.