manual: warn that --manual without hook wont renew
Sometimes users do not understand that obtaining a certificate using
`-manual` (without `--manual-auth-hook`) means that Certbot will not be
able to automatically renew that certificate. This results in an
unwelcome surprise at renewal time.

This change introduces a proactive yes/no prompt explaining the
situation and tries to point them toward either using a hook or using a
different plugin. It appears only during interactive execution.