Note required scopes for Linode API credentials
The documentation for certbot-dns-linode doesn't currently mention which Linode
API scopes are required for it to function properly, which may result in users either
over-granting permissions to the credentials they provide or encountering cryptic
errors when attempting to obtain a certificate. Fix that by providing details on the
which scopes are necessary for this plugin to function.

Also clean up a link to the old Linode manager tool which is no longer available
(and automatically redirects to the main page of the new Cloud console).