Raise error when an unknown `certificates` output is chosen
At the moment, the `certbot certificates` output style is internally
hardcoded and only changable with a flag, but we might want to change
that in the future by adding a single command line option where the
user can input the required output style as free text (e.g.
`--output-format yaml` in stead the currently  hardcoded option `--json`.
This commit already anticipates on such a change to check the output
format used in `_report_human_readable()` and raise an error if it's
an output it doesn't recognise.