BF: apache cfg parsing - relax assumption that value cannot contain =
I have

   Define: goto=http://datasets.datalad.org/?dir=/shub

which seems to be perfectly fine by apache.  But the = in URL throws
certbot off causing an error, forbidding certificates updates for any
of the sites etc.

It seems only logical to simply split once by a single (if present) =.

I have not looked/analyzed into the other spot of use of
.partition("=") in parse_define_file -- may be similar approach needs
to be implemented there as well.