Support read-only root
When using a root filesystem that's mounted read-only, certbot fails handle
the lock for its configuration in */etc/letsencrypt*. The easiest solution
is to place a symlink at */etc/letsencrypt/.certbot.lock* that points to a
writeable location like */var/lib/letsencrypt/etc-letsencrypt-lock*.

But even this requires some tweaking of the code: At release time, we have
to check for a symlink and don't remove it but the target.

To get a writeable filesystem during operation the filesystem must be
remounted rw. This could be done with a pre-hook, but they are run to late.
Hence, the run of the pre-hook is scheduled earlier to allow mounting the
filesystem read-write before any access takes place.