(Re-)Add the `--autorenew` option.
In #5911, initially the `--autorenew` option was present, however, it
was removed in cd065b3 again for some reason. That means that if the
user has used `--no-autorenew` in the past, there isn't any way to
disable the `autorenew = False` flag in the renewal configuration file.
This PR re-introduces the `--autorenew` flag, so the user can re-enable
autorenewal for certs where autorenewal was disabled previously.