Find zone based on validation name
If you've set up delegation of the _acme-challenge subdomain to a
validation-specific zone in Google Cloud DNS the plugin fails to
find the correct zone since the current lookup is based on the domain.

This changes the zone lookup for the Google Cloud DNS validator to
the validation_name instead of the domain argument. The validation_name
is usually just the domain name with the "_acme-challenge." prefix.